requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
pandas>=2.1.0
openpyxl>=3.1.0
transformers>=4.36.0
torch>=2.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
lxml>=4.9.0
newspaper3k>=0.1.0
nltk>=3.9.0